+++ offchain/crates/solana-admin-cli/revenue-distribution/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- update migration command ([#195](https://github.com/doublezerofoundation/doublezero-offchain/pull/195))
+
 ## [0.0.1](https://github.com/doublezerofoundation/doublezero-offchain/compare/doublezero-revenue-distribution-admin-cli/v0.0.0...doublezero-revenue-distribution-admin-cli/v0.0.1) - 2025-10-21
 
 ### Other